+++ docs/source/changelog.rst
@@ -6,6 +6,7 @@ Change Log
 
 - dropped support for Python 3.4 (which has reached end-of-life status). Supported Python versions are now 2.7, and 3.5 or newer.
   (the life cycle status of the Python versions can be seen here https://devguide.python.org/#status-of-python-branches)
+- URIs now allow spaces in the location part. Useful for unix domain sockets.
 
 
 **Pyro 4.76**